Fix failing test + refactoring monaco service
