--- packages/host/tests/integration/modifiers/monaco-test.js
@@ -1,15 +1,0 @@
-import { module, test } from 'qunit';
-import { setupRenderingTest } from 'ember-qunit';
-import { render } from '@ember/test-helpers';
-import { hbs } from 'ember-cli-htmlbars';
-
-module('Integration | Modifier | monaco', function (hooks) {
-  setupRenderingTest(hooks);
-
-  // Replace this with your real tests.
-  test('it renders', async function (assert) {
-    await render(hbs`<div {{monaco}}></div>`);
-
-    assert.ok(true);
-  });
-});
